The things which was finished last time have some bugs, so I fix them again. Now we are missing taking 0's latest time into consideration and improveTimeConsistency.

+++ Solution.h
@@ -170,8 +170,8 @@ protected:
 	void printTimeLine(const std::vector<PointAndType>& pointsAndTypes);
 	void printVerticalLines(int lines);
 
-	int getViolationScore(const std::vector<std::vector<int>>& solutionListOfEachDay, int scaleOfViolationScore);
-	int getObjectiveScore(const std::vector<std::vector<int>>& solutionListOfEachDay);
+	double getViolationScore(const std::vector<std::vector<int>>& solutionListOfEachDay, double scaleOfViolationScore);
+	double getObjectiveScore(const std::vector<std::vector<int>>& solutionListOfEachDay);
 };
 
 #endif
